sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dap-sdk"
version = "0.1.0"
description = "Python SDK for the Delegated Agent Protocol (DAP) v0.1"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "DAP Protocol", email = "dev@dap-protocol.org" },
]
keywords = ["dap", "agent", "protocol", "credential", "delegation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Typing :: Typed",
]
dependencies = [
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.urls]
Homepage = "https://github.com/dap-protocol/dap"
Specification = "https://github.com/dap-protocol/dap/blob/main/spec/v0.1.md"
